requirements.txt
discord.py
motor
colorama